增加商品分类接口

+++ src/main/java/net/chenlin/dp/modules/biz/bussiness/entity/CommodityEntity.java
@@ -49,6 +49,6 @@ public class CommodityEntity implements Serializable {
     private Timestamp modify_time;
 
     @ApiModelProperty(value = "机构ID")
-    private int org_id;
+    private Long org_id;
 
 }
